Local

+++ pets/serializers/pet_serializer.py
@@ -15,7 +15,7 @@ class PetSerializer(serializers.Serializer):
     coat_texture = serializers.CharField(max_length=200)
     coat_pattern = serializers.CharField(max_length=200)
     birth_date = serializers.DateTimeField()
-    weight = serializers.IntegerField()
+    weight = serializers.DecimalField(max_digits=5, decimal_places=2)
     specie = serializers.CharField(max_length=200)
     sex = serializers.CharField(max_length=200)
     is_castrated = serializers.BooleanField()
